- SDK updated. - Delete app's own mechanism for extensions. - Make use of the new extension APIs supporting multiple location identifiers. - Place the static init methods for each extension in its own class instead of an OCExtension extension.
