+++ ownCloud/SDK Extensions/OCExtensionType+Extension.swift
@@ -1,0 +1,13 @@
+//
+//  OCExtension+DisplayView.swift
+//  ownCloud
+//
+//  Created by Pablo Carrascal on 30/08/2018.
+//  Copyright © 2018 ownCloud GmbH. All rights reserved.
+//
+
+import ownCloudSDK
+
+extension OCExtensionType {
+	static let viewer: OCExtensionType  =  OCExtensionType(rawValue: "viewer")
+}
