+++ ownCloud/Viewer/PDFViewerViewController.swift
@@ -13,7 +13,19 @@ import PDFKit
 class PDFViewerViewController: DisplayViewController, DisplayViewProtocol {
 
 	static var supportedMimeTypes: [String] = ["application/pdf"]
-	static var features: [String : Any]? = [FeatureKeys.canEdit : true, FeatureKeys.showPDF : true]
+	static var features: [String : Any]? = [FeatureKeys.canEdit : true]
+
+	static func normalPDFExtension(identifier: String) -> OCExtension {
+		let rawIdentifier: OCExtensionIdentifier =  OCExtensionIdentifier(rawValue: identifier)
+		let locationIdentifier = OCExtensionLocationIdentifier(rawValue: supportedMimeTypes[0])
+		let features: [String : Any] = PDFViewerViewController.features!
+
+		let normalPDFExtension = OCExtension(identifier: rawIdentifier, type: .viewer, location: locationIdentifier, features: features) { (_ rawExtension, _ context, _ error) -> Any? in
+			return PDFViewerViewController()
+		}
+
+		return normalPDFExtension!
+	}
 
 	override func renderSpecificView() {
 		if let document = PDFDocument(url: source) {
